requirement.txt
aiofiles==0.7.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
dataclasses==0.8
demjson==2.2.4
fastapi==0.61.0
future==0.18.2
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
importlib-resources==5.1.4
netmiko==3.4.0
ntc-templates==2.0.0
numpy==1.19.5
paramiko==2.7.2
psutil==5.8.0
pycparser==2.20
pydantic==1.6.1
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyroute2==0.5.19
pyserial==3.5
python-multipart==0.0.5
redis==3.5.3
requests==2.20.0
SQLAlchemy==1.3.19
starlette==0.13.6
setproctitle==1.2.2
tenacity==7.0.0
textfsm==1.1.0
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.14.0
websockify==0.9.0
Werkzeug==2.0.0
zipp==3.4.1
